Mods

+++ upload.php
@@ -221,11 +221,15 @@
     global $paths;
 
     for ($i = 0; $i < count($paths); $i++) {
-      unlink($paths[$i]);
+        if (!empty($paths[$i]) && file_exists($paths[$i])) {
+            unlink($paths[$i]);
+        }
     }
 
     for ($i = 0; $i < count($uploaded); $i++) {
-      unlink($uploaded[$i]);
+        if (!empty($uploaded[$i]) && file_exists($uploaded[$i])) {
+            unlink($uploaded[$i]);
+        }
     }
   }
 ?>
